pyproject.toml
[tool.poetry]
name = "streaming"
version = "0.1.0"
description = ""
authors = ["Giovanny Baron <obaron@mabtec.com.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
